% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharksay.R
\name{shark_upper}
\alias{shark_upper}
\title{Make the shark shout in uppercase}
\usage{
shark_upper(message = "shouting!")
}
\arguments{
\item{message}{A character string. The message you want the shark to shout.}
}
\value{
No return value. This function outputs text to the console.
}
\description{
Displays a message in uppercase with a shark (ASCII art).
}
\examples{
shark_upper("i am loud!")

}
