% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharksay.R
\name{sharksay}
\alias{sharksay}
\title{Make the shark say something}
\usage{
sharksay(message = "I am a shark!")
}
\arguments{
\item{message}{A character string. The message you want the shark to say.}
}
\value{
No return value. This function outputs text to the console.
}
\description{
Displays a message with a shark (ASCII art) in the R console.
}
\examples{
sharksay("Hello, I am a shark!")

}
